Fix html generation

+++ src/Event/Leaderboard/Entry.php
@@ -35,7 +35,7 @@ final class Entry implements \Stringable
         assert(count($data) === count(self::HEADER));
 
         return new self(
-            new Hero($data[0], $data[1], $data[2], strtr($data[3], ',', '.')),
+            new Hero($data[0], $data[1], $data[2], (float) strtr($data[3], ',', '.')),
             new Trial($data[4], $data[5], $data[6], $data[7], $data[8], $data[9]),
             $data[10],
         );
